DBZ-6746 Correctly manage precision and scale with numeric types

+++ src/test/java/io/debezium/connector/jdbc/e2e/JdbcSinkPipelineToMySqlIT.java
@@ -63,7 +63,7 @@ public class JdbcSinkPipelineToMySqlIT extends AbstractJdbcSinkPipelineIT {
 
     @Override
     protected String getVariableScaleDecimalType() {
-        return "DOUBLE";
+        return "DECIMAL";
     }
 
     @Override
