+++ src/test/java/io/debezium/connector/jdbc/e2e/JdbcSinkPipelineToSqlServerIT.java
@@ -98,7 +98,7 @@ public class JdbcSinkPipelineToSqlServerIT extends AbstractJdbcSinkPipelineIT {
 
     @Override
     protected String getVariableScaleDecimalType() {
-        return "FLOAT";
+        return "NUMERIC";
     }
 
     @Override
